Splitting model into small files. Reusing code between literals
