+++ src/model/node.js
@@ -1,0 +1,1 @@
+export const Node = nodeType => body => ({ nodeType, ...body })
